implemento headers para datatypes. Falta realizar validaciones de parámetros

+++ include/datatypes/DTNormal.h
@@ -9,7 +9,7 @@ class DTNormal: public DTDistribucion {
         float sigmaSquare;
     public:
         DTNormal(float,float);
-        int getMu() const;
+        float getMu() const;
         float getSigmaSquare() const;
 };
